Extract tests for isRangeError to own file
